LPS-54341 Rename
