+++ modules/apps/blogs/blogs-editor-config/.project
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
 <projectDescription>
-	<name>blogs-portlet-editor-config-master</name>
+	<name>blogs-editor-config-master</name>
 	<comment></comment>
 	<projects></projects>
 	<buildSpec>
