+++ modules/apps/blogs/blogs-editor-config/bnd.bnd
@@ -1,0 +1,4 @@
+Bundle-Name: Liferay Blogs Editor Config
+Bundle-SymbolicName: com.liferay.blogs.editor.config
+Bundle-Version: 1.0.0
+Include-Resource: classes
