+++ modules/apps/blogs/blogs-editor-config/build.xml
