+++ modules/apps/blogs/blogs-editor-config/src/com/liferay/blogs/editor/config/BlogsContentEditorConfigContributor.java
@@ -12,17 +12,16 @@
  * details.
  */
 
-package com.liferay.blogs.portlet.editor.config;
+package com.liferay.blogs.editor.config;
 
 import com.liferay.portal.kernel.editor.config.EditorConfigContributor;
-import com.liferay.portal.kernel.json.JSONArray;
-import com.liferay.portal.kernel.json.JSONException;
-import com.liferay.portal.kernel.json.JSONFactoryUtil;
 import com.liferay.portal.kernel.json.JSONObject;
 import com.liferay.portal.kernel.portlet.LiferayPortletResponse;
 import com.liferay.portal.theme.ThemeDisplay;
 
 import java.util.Map;
+
+import javax.portlet.PortletURL;
 
 import org.osgi.service.component.annotations.Component;
 
@@ -32,32 +31,16 @@ import org.osgi.service.component.annotations.Component;
 @Component(
 	property = {
 		"javax.portlet.name=33", "javax.portlet.name=161",
-		"editor.config.key=coverImageCaptionEditor"
+		"editor.config.key=contentEditor"
 	}
 )
-public class BlogsCoverImageCaptionEditorConfigContributor
+public class BlogsContentEditorConfigContributor
 	implements EditorConfigContributor {
 
 	public void populateConfigJSONObject(
 		JSONObject jsonObject, Map<String, Object> inputEditorTaglibAttributes,
 		ThemeDisplay themeDisplay,
 		LiferayPortletResponse liferayPortletResponse) {
-
-		jsonObject.put("allowedContent", "a");
-		jsonObject.put("disallowedContent", "br");
-
-		JSONObject toolbarJSONObject = JSONFactoryUtil.createJSONObject();
-
-		try {
-			JSONArray stylesJSONObject = JSONFactoryUtil.createJSONArray(
-				"['a']");
-
-			toolbarJSONObject.put("styles", stylesJSONObject);
-		}
-		catch (JSONException jsone) {
-		}
-
-		jsonObject.put("toolbars", toolbarJSONObject);
 	}
 
 	public void populateOptionsJSONObject(
@@ -65,7 +48,13 @@ public class BlogsCoverImageCaptionEditorConfigContributor
 		ThemeDisplay themeDisplay,
 		LiferayPortletResponse liferayPortletResponse) {
 
-		jsonObject.put("textMode", Boolean.TRUE);
+		PortletURL uploadEditorImageURL =
+			liferayPortletResponse.createActionURL();
+
+		uploadEditorImageURL.setParameter(
+			"struts_action", "/blogs/upload_editor_image");
+
+		jsonObject.put("uploadURL", uploadEditorImageURL.toString());
 	}
 
 }
