+++ modules/apps/blogs/blogs-editor-config/src/com/liferay/blogs/editor/config/BlogsCoverImageCaptionEditorConfigContributor.java
@@ -12,9 +12,11 @@
  * details.
  */
 
-package com.liferay.blogs.portlet.editor.config;
+package com.liferay.blogs.editor.config;
 
 import com.liferay.portal.kernel.editor.config.EditorConfigContributor;
+import com.liferay.portal.kernel.json.JSONArray;
+import com.liferay.portal.kernel.json.JSONException;
 import com.liferay.portal.kernel.json.JSONFactoryUtil;
 import com.liferay.portal.kernel.json.JSONObject;
 import com.liferay.portal.kernel.portlet.LiferayPortletResponse;
@@ -30,11 +32,10 @@ import org.osgi.service.component.annotations.Component;
 @Component(
 	property = {
 		"javax.portlet.name=33", "javax.portlet.name=161",
-		"editor.config.key=descriptionEditor",
-		"editor.config.key=subtitleEditor", "editor.config.key=titleEditor"
+		"editor.config.key=coverImageCaptionEditor"
 	}
 )
-public class BlogsTextEditorConfigContributor
+public class BlogsCoverImageCaptionEditorConfigContributor
 	implements EditorConfigContributor {
 
 	public void populateConfigJSONObject(
@@ -42,9 +43,21 @@ public class BlogsTextEditorConfigContributor
 		ThemeDisplay themeDisplay,
 		LiferayPortletResponse liferayPortletResponse) {
 
-		jsonObject.put("allowedContent", "p");
+		jsonObject.put("allowedContent", "a");
 		jsonObject.put("disallowedContent", "br");
-		jsonObject.put("toolbars", JSONFactoryUtil.createJSONObject());
+
+		JSONObject toolbarJSONObject = JSONFactoryUtil.createJSONObject();
+
+		try {
+			JSONArray stylesJSONObject = JSONFactoryUtil.createJSONArray(
+				"['a']");
+
+			toolbarJSONObject.put("styles", stylesJSONObject);
+		}
+		catch (JSONException jsone) {
+		}
+
+		jsonObject.put("toolbars", toolbarJSONObject);
 	}
 
 	public void populateOptionsJSONObject(
