--- modules/apps/blogs/blogs-portlet-editor-config/bnd.bnd
@@ -1,4 +1,0 @@
-Bundle-Name: Liferay Blogs Portlet Editor Config
-Bundle-SymbolicName: com.liferay.blogs.portlet.editor.config
-Bundle-Version: 1.0.0
-Include-Resource: classes
